requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
typer==0.19.1
typing_extensions==4.15.0
urllib3==2.5.0
